backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
python-multipart>=0.0.18
chatterbox-tts>=0.1.4
torch>=2.0
torchaudio>=2.0
pymupdf>=1.24
python-docx>=1.1
pytesseract>=0.3.10
Pillow>=10.0
pydub>=0.25
langdetect>=1.0.9
python-dotenv>=1.0
